Sekiban.Web.OpenApi: Refactor OpenApiFilter and CustomSchemaIdSelector.
