+++ src/Sekiban.Web/OpenApi/Extensions/IServiceCollectionExtensions.cs
@@ -1,20 +1,22 @@
 ﻿using Microsoft.Extensions.DependencyInjection;
 using Sekiban.Web.Dependency;
 using Swashbuckle.AspNetCore.SwaggerGen;
-
 namespace Sekiban.Web.OpenApi.Extensions;
+
 public static class IServiceCollectionExtensions
 {
     public static IServiceCollection AddSwaggerGenWithSekibanOpenApiFilter(this IServiceCollection services,
-        Action<SwaggerGenOptions>? setupAction = null)
+        Action<SwaggerGenOptions>? setupSwaggerGenAction = null,
+        Func<Type, string>? openApiCustomSchemaIdSelector = null)
     {
         services.AddSwaggerGen(options =>
         {
-            options.CustomSchemaIds(type => SekibanOpenApiParameterGenerator.GenerateCustomSchemaName(type));
+            options.CustomSchemaIds(openApiCustomSchemaIdSelector ?? SekibanOpenApiParameterGenerator.GenerateCustomSchemaName);
+
             options.SchemaFilter<SekibanOpenApiFilter>();
             options.OperationFilter<SekibanOpenApiFilter>();
 
-            setupAction?.Invoke(options);
+            setupSwaggerGenAction?.Invoke(options);
         });
 
         return services;
