+++ src/Sekiban.Web/OpenApi/SekibanOpenApiFilter.cs
@@ -1,11 +1,14 @@
-﻿using Microsoft.OpenApi.Models;
+﻿using Microsoft.OpenApi.Any;
+using Microsoft.OpenApi.Models;
 using Sekiban.Web.OpenApi.Extensions;
 using Swashbuckle.AspNetCore.SwaggerGen;
+using System.ComponentModel;
+using System.ComponentModel.DataAnnotations;
+using System.Reflection;
 namespace Sekiban.Web.OpenApi;
 
 public class SekibanOpenApiFilter : ISchemaFilter, IOperationFilter
 {
-
     public void Apply(OpenApiOperation operation, OperationFilterContext context)
     {
         if (operation.RequestBody is not null && (operation.RequestBody.Content is null || !operation.RequestBody.Content.Any()))
@@ -13,11 +16,12 @@ public class SekibanOpenApiFilter : ISchemaFilter, IOperationFilter
             operation.RequestBody = null;
         }
     }
+
     public void Apply(OpenApiSchema schema, SchemaFilterContext context)
     {
         if (context.Type is not null && context.Type.IsEnum)
         {
-            SekibanOpenApiParameterGenerator.GenerateSchemaForEnum(context.Type, schema);
+            GenerateSchemaForEnum(context.Type, schema);
         }
 
         (schema.Title, schema.Description) = context switch
@@ -33,5 +37,48 @@ public class SekibanOpenApiFilter : ISchemaFilter, IOperationFilter
 
             _ => (schema.Title, schema.Description)
         };
+    }
+
+    private static OpenApiSchema GenerateSchemaForEnum(Type propertyType, OpenApiSchema? schema = default)
+    {
+        var baseType = Nullable.GetUnderlyingType(propertyType) ?? propertyType;
+
+        schema ??= new OpenApiSchema();
+        schema.Type = baseType.Name;
+        schema.Nullable = Nullable.GetUnderlyingType(propertyType) is not null;
+
+        var enums = new OpenApiArray();
+        enums.AddRange(Enum.GetValues(baseType).Cast<Enum>().Select(enm => new OpenApiString(enm.ToString())));
+        schema.Enum = enums;
+
+        var displayNames = Enum.GetValues(baseType)
+            .Cast<Enum>()
+            .Select(
+                enm => enm.GetType().GetMember(enm.ToString()) is { } members && members.FirstOrDefault() is { } member
+                    ? member.GetCustomAttribute<DisplayAttribute>()?.Name ?? member.GetCustomAttribute<DisplayNameAttribute>()?.DisplayName
+                    : null)
+            .ToList();
+        if (displayNames.Any(a => !string.IsNullOrEmpty(a)))
+        {
+            var enumVarNames = new OpenApiArray();
+            enumVarNames.AddRange(displayNames.Select(s => new OpenApiString(s)));
+            schema.Extensions.Add("x-enum-varnames", enumVarNames);
+        }
+
+        var descriptions = Enum.GetValues(baseType)
+            .Cast<Enum>()
+            .Select(
+                enm => enm.GetType().GetMember(enm.ToString()) is { } members && members.FirstOrDefault() is { } member
+                    ? member.GetCustomAttribute<DisplayAttribute>()?.Description ?? member.GetCustomAttribute<DescriptionAttribute>()?.Description
+                    : null)
+            .ToList();
+        if (descriptions.Any(a => !string.IsNullOrEmpty(a)))
+        {
+            var enumDescriptions = new OpenApiArray();
+            enumDescriptions.AddRange(descriptions.Select(s => new OpenApiString(s)));
+            schema.Extensions.Add("x-enum-descriptions", enumDescriptions);
+        }
+
+        return schema;
     }
 }
